update 3d plot settings
I think it's a little better to floor rather than round the number of rows (e.g. 7 and 8 views seems to me better as 2 rather than 3 rows). This matters for me adding a 3d plot to 6 views for example.

I modified the default axis limits for the 3d plot to use defScale when it's defined by user applicable.